requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.2.0
